refactor: extract WorkspaceState new-chat and navigation extensions
Move the new-chat composer (recipient lookup, validation, creation) into
WorkspaceState+NewChat.swift and the top-level navigation actions (chat
selection, settings/new-chat presentation, chat-list toggle) into
WorkspaceState+Navigation.swift. Pure code movement, no behavior change.
Build green; all 208 tests pass.
